refactor: ChatService와 ChatSessionService 코드 정리
